# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: A robust microservice backend for managing teams, users, and digital
    assets.
  termsOfService: http://swagger.io/terms/
  title: Team Management API
  version: "1.0"
paths:
  /api/teams/:
    post:
      consumes:
      - application/json
      description: Allows a manager to create a new team
      parameters:
      - description: 'Team Name (e.g., {''name'': ''My Team''})'
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new team
      tags:
      - Teams
  /api/teams/{id}:
    delete:
      consumes:
      - application/json
      description: Allows a manager to delete a team
      parameters:
      - description: Team ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a team
      tags:
      - Teams
    get:
      consumes:
      - application/json
      description: Retrieve a team by its unique identifier
      parameters:
      - description: Team ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get team by ID
      tags:
      - Teams
    put:
      consumes:
      - application/json
      description: Allows a manager to update a team's information
      parameters:
      - description: Team ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 'Updated Team Name (e.g., {''name'': ''Updated Team Name''})'
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a team
      tags:
      - Teams
  /api/teams/{id}/members:
    post:
      consumes:
      - application/json
      description: Allows a manager to add a user to a team
      parameters:
      - description: Team ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 'User ID (e.g., {''user_id'': 123})'
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            format: int64
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add member to team
      tags:
      - Teams
  /api/teams/users/{userID}:
    get:
      consumes:
      - application/json
      description: Retrieve all teams associated with a specific user
      parameters:
      - description: User ID
        format: int64
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get teams by user ID
      tags:
      - Teams
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"